Cleanup, nits, (C)
