+++ src/Microsoft.Extensions.Logging.AzureAppServices/Internal/BatchingLogger.cs
@@ -1,0 +1,59 @@
+﻿// Copyright (c) .NET Foundation. All rights reserved.
+// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
+
+using System;
+using System.Text;
+
+namespace Microsoft.Extensions.Logging.AzureAppServices.Internal
+{
+    public class BatchingLogger : ILogger
+    {
+        private readonly BatchingLoggerProvider _provider;
+        private readonly string _category;
+
+        public BatchingLogger(BatchingLoggerProvider loggerProvider, string categoryName)
+        {
+            _provider = loggerProvider;
+            _category = categoryName;
+        }
+
+        public IDisposable BeginScope<TState>(TState state)
+        {
+            return null;
+        }
+
+        public bool IsEnabled(LogLevel logLevel)
+        {
+            return _provider.IsEnabled;
+        }
+
+        public void Log<TState>(DateTimeOffset timestamp, LogLevel logLevel, EventId eventId, TState state, Exception exception, Func<TState, Exception, string> formatter)
+        {
+            if (!IsEnabled(logLevel))
+            {
+                return;
+            }
+
+            var builder = new StringBuilder();
+            builder.Append(timestamp.ToString("yyyy-MM-dd HH:mm:ss.fff zzz"));
+            builder.Append(" [");
+            builder.Append(logLevel.ToString());
+            builder.Append("] ");
+            builder.Append(_category);
+            builder.Append(": ");
+            builder.AppendLine(formatter(state, exception));
+
+            if (exception != null)
+            {
+                builder.AppendLine(exception.ToString());
+            }
+
+            _provider.AddMessage(timestamp, builder.ToString());
+        }
+
+        public void Log<TState>(LogLevel logLevel, EventId eventId, TState state, Exception exception, Func<TState, Exception, string> formatter)
+        {
+            Log(DateTimeOffset.Now, logLevel, eventId, state, exception, formatter);
+        }
+    }
+}
