+++ src/Microsoft.Extensions.Logging.AzureAppServices/Internal/ConfigurationBasedLevelSwitcher.cs
@@ -1,4 +1,7 @@
-﻿using System;
+﻿// Copyright (c) .NET Foundation. All rights reserved.
+// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
+
+using System;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.Options;
