+++ src/Microsoft.Extensions.Logging.AzureAppServices/Internal/LogMessage.cs
@@ -1,0 +1,13 @@
+﻿// Copyright (c) .NET Foundation. All rights reserved.
+// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
+
+using System;
+
+namespace Microsoft.Extensions.Logging.AzureAppServices.Internal
+{
+    public struct LogMessage
+    {
+        public DateTimeOffset Timestamp { get; set; }
+        public string Message { get; set; }
+    }
+}
