+++ src/Microsoft.Extensions.Logging.AzureAppServices/Internal/SiteConfigurationProvider.cs
@@ -1,9 +1,12 @@
-﻿using System.IO;
+﻿// Copyright (c) .NET Foundation. All rights reserved.
+// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
+
+using System.IO;
 using Microsoft.Extensions.Configuration;
 
 namespace Microsoft.Extensions.Logging.AzureAppServices.Internal
 {
-    public class AzureDiagnosticsConfigurationProvider
+    public class SiteConfigurationProvider
     {
         public static IConfiguration GetAzureLoggingConfiguration(IWebAppContext context)
         {
