+++ test/Microsoft.Extensions.Logging.AzureAppServices.Test/AzureDiagnosticsConfigurationProviderTests.cs
@@ -20,7 +20,7 @@ namespace Microsoft.Extensions.Logging.AzureAppServices.Test
             contextMock.SetupGet(c => c.HomeFolder)
                 .Returns(tempFolder);
 
-            var config = AzureDiagnosticsConfigurationProvider.GetAzureLoggingConfiguration(contextMock.Object);
+            var config = SiteConfigurationProvider.GetAzureLoggingConfiguration(contextMock.Object);
 
             Assert.NotNull(config);
         }
@@ -46,7 +46,7 @@ namespace Microsoft.Extensions.Logging.AzureAppServices.Test
                 contextMock.SetupGet(c => c.HomeFolder)
                     .Returns(tempFolder);
 
-                var config = AzureDiagnosticsConfigurationProvider.GetAzureLoggingConfiguration(contextMock.Object);
+                var config = SiteConfigurationProvider.GetAzureLoggingConfiguration(contextMock.Object);
 
                 Assert.Equal("test value", config["key"]);
                 Assert.Equal("USEFUL_VALUE", config["RANDOM_ENVIRONMENT_VARIABLE"]);
