+++ test/Microsoft.Extensions.Logging.AzureAppServices.Test/BatchingLoggerProviderTests.cs
@@ -1,8 +1,12 @@
-﻿using System;
+﻿// Copyright (c) .NET Foundation. All rights reserved.
+// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
+
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
+using Microsoft.Extensions.Logging.AzureAppServices.Internal;
 using Xunit;
 
 namespace Microsoft.Extensions.Logging.AzureAppServices.Test
