+++ test/Microsoft.Extensions.Logging.AzureAppServices.Test/FileLoggerTests.cs
@@ -1,7 +1,11 @@
+// Copyright (c) .NET Foundation. All rights reserved.
+// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
+
 using System;
 using System.IO;
 using System.Linq;
 using System.Threading.Tasks;
+using Microsoft.Extensions.Logging.AzureAppServices.Internal;
 using Xunit;
 
 namespace Microsoft.Extensions.Logging.AzureAppServices.Test
