+++ test/Microsoft.Extensions.Logging.AzureAppServices.Test/OptionsWrapperMonitor.cs
@@ -1,3 +1,6 @@
+// Copyright (c) .NET Foundation. All rights reserved.
+// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
+
 using System;
 using Microsoft.Extensions.Options;
